moving the Meta and CachedMeta classes into their own directory to avoid a bunch of circular imports when pulling them from sagworks.api
