+++ examples/meta/meta_model_metrics.py
@@ -1,4 +1,4 @@
-from sageworks.api.meta import Meta
+from sageworks.meta import Meta
 
 # Create our Meta Class and get a list of our Models
 meta = Meta()
