+++ examples/plugins/pages/plugin_page_2.py
@@ -6,7 +6,7 @@ import dash_bootstrap_components as dbc
 
 # SageWorks Imports
 from sageworks.web_components import table
-from sageworks.api.meta import Meta
+from sageworks.meta import Meta
 
 
 class PluginPage2:
