+++ src/sageworks/api/__init__.py
@@ -16,8 +16,6 @@ from .data_source import DataSource
 from .feature_set import FeatureSet
 from .model import Model, ModelType
 from .endpoint import Endpoint
-from .meta import Meta
-from .cached_meta import CachedMeta
 from .parameter_store import ParameterStore
 from .df_store import DFStore
 
@@ -27,8 +25,6 @@ __all__ = [
     "Model",
     "ModelType",
     "Endpoint",
-    "Meta",
-    "CachedMeta",
     "ParameterStore",
     "DFStore",
 ]
