+++ src/sageworks/plugins/pages/mdq_page.py
@@ -8,7 +8,7 @@ import dash_bootstrap_components as dbc
 # SageWorks Imports
 from sageworks.api import FeatureSet
 from sageworks.web_components import table
-from sageworks.api.meta import Meta
+from sageworks.meta import Meta
 from sageworks.web_components.plugins.data_details import DataDetails
 from sageworks.web_components.plugins.scatter_plot import ScatterPlot
